+++ common/d_scan.c
@@ -21,6 +21,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //
 // Portable C scan-level rasterization code, all pixel depths.
 
+#include <stdint.h>
+
 #include "quakedef.h"
 #include "r_local.h"
 #include "d_local.h"
@@ -893,7 +895,7 @@ D_DrawZSpans(espan_t *pspan)
       // we count on FP exceptions being turned off to avoid range problems
       int   izi = (int)(zi * 0x8000 * 0x10000);
 
-      if ((long)pdest & 0x02)
+      if ((int32_t)pdest & 0x02)
       {
          *pdest++ = (short)(izi >> 16);
          izi += izistep;
